Move files to single cargo project and create utils folder to remove duplicated code
