+++ .gitignore
@@ -5,6 +5,7 @@
 
 # Cargo
 **/target
+/target
 
 # Mac
 .DS_Store
