+++ Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "decen"
+name = "parallel-astar-rust"
 version = "0.1.0"
 authors = ["Brett Fazio, Ellie Kozlowski, Dylan Ochoa, Blake Robertson, Idel Martinez"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 crossbeam = "0.8"
+rand = "0.8.1"
+clap = "2.27.0"
