--- hda/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "hda"
-version = "0.1.0"
-authors = ["Dylan Ochoa <dylan.ochoa2323@gmail.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-crossbeam = "0.8"
-rand = "0.8.1"
-clap = "2.27.0"
