+++ src/a_star/utils/mod.rs
@@ -1,0 +1,3 @@
+pub mod dynamic_barrier;
+pub mod helpers;
+pub mod structs;
